refactor : throttlingManager 패키지명 변경(#72)
